[feat] Add Coordinator Protocol and BaseCoordinator.swift
