--- GeulGeurim/Shared/Coordinator.swift
@@ -1,37 +1,0 @@
-//
-//  Coordinator.swift
-//  GeulGeurim
-//
-//  Created by HUNHEE LEE on 28.05.2024.
-//
-
-import UIKit
-
-public protocol CoordinatorFinishDelegate: AnyObject {
-  func coordinatorDidFinish(childCoordinator: Coordinator)
-}
-
-public protocol Coordinator: AnyObject {
-  var finishDelegate: CoordinatorFinishDelegate? { get set }
-  var navigationController: UINavigationController { get set }
-  var childCoordinators: [Coordinator] { get set }
-  
-  func start()
-  func finish()
-}
-
-public extension Coordinator {
-  func finish() {
-    childCoordinators.removeAll()
-    finishDelegate?.coordinatorDidFinish(childCoordinator: self)
-  }
-  
-  func removeChildCoordinator(_ childCoordinator: Coordinator) {
-    for (index, coordinator) in childCoordinators.enumerated() {
-      if coordinator === childCoordinator {
-        childCoordinators.remove(at: index)
-        break
-      }
-    }
-  }
-}
