+++ GeulGeurim/Shared/DesignSystem/Sources/Base/BaseView.swift
@@ -1,0 +1,8 @@
+//
+//  BaseView.swift
+//  GeulGeurim
+//
+//  Created by HUNHEE LEE on 1.06.2024.
+//
+
+import Foundation
